requirements.txt
streamlit==1.28.1
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
openai==1.3.7
python-dotenv==1.0.0
pandas==2.1.3
plotly==5.17.0
reportlab==4.0.7
markdown==3.5.1
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
fake-useragent==1.4.0
trafilatura==1.6.4
readability-lxml==0.8.1
langdetect==1.0.9
textstat==0.7.3
wordcloud==1.9.2
matplotlib==3.8.2
seaborn==0.13.0
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
chromadb==0.4.15
faiss-cpu==1.7.4
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-learn==1.3.2
numpy==1.24.3
aiohttp==3.9.1
anthropic==0.3.0
